Removed AbstractILikeStrFunctionFactory.java
Removed AbstractILikeStrFunctionFactory class and implemented implemented the flag in AbstractLikeStrFunctionFactory class
